implement occupancy calculation for non-report endpoints
